+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.18.0.0
 
+* Rename `wvkSig` to `wvkSignature`
 * Remove `eqBootstrapWitnessRaw` and `BootstrapWitnessRaw`
 * Rename `bwSig` to `bwSignature` for `BootstrapWitness`
 * Remove `witVKeyBytes` and `eqWitVKeyRaw`